search_agent_RAG/app/__main__.py
import logging
import os
import sys

import httpx
import uvicorn
from a2a.server.apps import A2AStarletteApplication
from a2a.server.request_handlers import DefaultRequestHandler
from a2a.server.tasks import InMemoryPushNotifier, InMemoryTaskStore
from a2a.types import (
    AgentCapabilities,
    AgentCard,
    AgentSkill,
)
from app.agent import search_agent_RAG
from app.agent_executor import search_agent_RAGExecutor
from dotenv import load_dotenv

load_dotenv()

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class MissingAPIKeyError(Exception):
    """Exception for missing API key."""


def main():
    """Starts search_agent_RAG server."""
    host = "localhost"
    port = 10004
    try:
        if not os.getenv("GOOGLE_API_KEY"):
            raise MissingAPIKeyError("GOOGLE_API_KEY environment variable not set.")

        capabilities = AgentCapabilities(streaming=True, pushNotifications=True)
        skill = AgentSkill(
            id="best_candidates_from_resumes",
            name="Candidates matching Tool",
            description="Helps with finding matching candidates based on job description",
            tags=["candidates", "matching", "job description"],
            examples=["IT Support Technician", "Software Developer", "Data Scientist"],
        )
        agent_card = AgentCard(
            name="search_agent_RAG",
            description="Specialized assistant to find matching candidates based on job description", 
            url=f"http://{host}:{port}/",
            version="1.0.0",
            defaultInputModes=search_agent_RAG.SUPPORTED_CONTENT_TYPES,
            defaultOutputModes=search_agent_RAG.SUPPORTED_CONTENT_TYPES,
            capabilities=capabilities,
            skills=[skill],
        )

        httpx_client = httpx.AsyncClient()
        request_handler = DefaultRequestHandler(
            agent_executor=search_agent_RAGExecutor(),
            task_store=InMemoryTaskStore(),
            push_notifier=InMemoryPushNotifier(httpx_client),
        )
        server = A2AStarletteApplication(
            agent_card=agent_card, http_handler=request_handler
        )

        uvicorn.run(server.build(), host=host, port=port)

    except MissingAPIKeyError as e:
        logger.error(f"Error: {e}")
        sys.exit(1)
    except Exception as e:
        logger.error(f"An error occurred during server startup: {e}")
        sys.exit(1)


if __name__ == "__main__":
    main()
